`break` out of decoder loop if a sequence jump is detected
the new `ignore_errors` flag can be used to show that there's not enough
data in the buffer to bother trying addtional decodes (br.size() == 1 in
all of my experiments) so `continue` wouldn't be better.

Change the exception raised by `br.unpack()` to `ValueError` if it's
called with an undersized buffer. That's an inappropriate value to be
unpacked and means I don't have to catch generic `Exception`.