more consistency checks and type enforcement in batch_read_vsfrs
Callers of batch_read_vsfrs should know the data type of each requested
VSFR, and should therefore be able to specify a correct format string.
Using an inappropriate format string will raise an error.

Require that vsfr_ids is a list of VSFR dataclasses. This prevents the use
of arbitrary integers as a VSFR to read. Using other data types will raise
an error.

Add a check for valid reads. The first item returned by the device is a
bitfield indicating which VSFRs were read successfully. Use this to check
that batch read operation was completely successful. As only the defined
VSFRs are permitted, and those were retrieved from the device, there should
be no reason for any read to fail as it might by retrieving an arbitrary
location (eg. 0xABAD1DEA). If such a failure is detected, an execption is
raised.