Decode more types
My RC102 needs a new battery; I may as well be mean to the old battery until
the new one arrives by subjecting it to extreme discharge and unreasonable
temperature excursions to capture those events.

Also, I had the opportunity to blast the sensor with a high dose, and was
able to capture the off-scale events. Sadly, I did not gain superpowers.